+++ tsconfig.json
@@ -37,7 +37,7 @@
     "paths": {
       "*": [
         "node_modules/*",
-        "src/types/*"
+        "src/resources/types/*"
       ]
     },
     "emitDecoratorMetadata": true,
